requirements.txt
jupyter==1.0.0
kafka-python==2.0.2
findspark==2.0.1
tweepy==4.6.0
python-dotenv==0.19.2